add: report completed and docker containers config added

+++ tech-test-spring/src/main/java/com/nttdata/alpaca/tech_test_spring/infraestructure/dto/MovementEventDTO.java
@@ -1,7 +1,5 @@
 package com.nttdata.alpaca.tech_test_spring.infraestructure.dto;
 
-import java.math.BigDecimal;
-
 import lombok.Data;
 
 @Data
@@ -10,8 +8,8 @@ public class MovementEventDTO {
     private String cliente;
     private String numeroCuenta;
     private String tipo;
-    private BigDecimal saldoInicial;
-    private BigDecimal valor;
-    private BigDecimal saldo;
+    private Double saldoInicial;
+    private Double valor;
+    private Double saldo;
     private String tipoMovimiento;
 }
